feat: remove explicit bot + migerate to latest

+++ schema/src/main/java/com/yk/schema/User.java
@@ -1,21 +1,13 @@
 package com.yk.schema;
 
-import java.util.HashSet;
-import java.util.Set;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotEmpty;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+
+import java.util.Set;
 
 @Builder
 @Getter
@@ -31,7 +23,7 @@ public class User {
     @Column(name = "id")
     private int id;
     @Column(name = "user_id")
-    private int userId;
+    private long userId;
     @OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
     Set<UserTopicAssignment> subMenu;
 }
